Refactor the constants
