--- compiler-plugin/src/main/java/io/ballerina/lib/np/compilerplugin/Constants.java
@@ -1,20 +1,0 @@
-package io.ballerina.lib.np.compilerplugin;
-
-public final class Constants {
-    public static final String NP = "np";
-    public static final String SCHEMA_ANNOTATION_IDENTIFIER = "schemaAnnot";
-    public static final String SCHEMA = "schema";
-    static final String ORG_NAME = "ballerinax";
-    static final String MODULE_NAME = "np";
-    static final String PROMPT_VAR = "prompt";
-    static final String MODEL_VAR = "model";
-    static final String LLM_CALL_ANNOT = "LlmCall";
-
-    static final String PROMPT_TYPE = "Prompt";
-    static final String MODEL_TYPE = "Model";
-    static final String CALL_LLM = "callLlm";
-    static final String COLON = ":";
-    static final String STRING = "string";
-    static final String BYTE = "byte";
-    static final String NUMBER = "number";
-}
